+++ Dayana/Shared/Persistence/Models/Identity/Base/ClaimBaseFile.cs
@@ -1,6 +1,11 @@
 ﻿using Dayana.Shared.Persistence.Models.Identity.Base.Users;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace Dayana.Shared.Persistence.Models.Identity.Base.Claims;
+namespace Dayana.Shared.Persistence.Models.Identity.Base;
 
 public class ClaimModel
 {
