+++ Dayana/Shared/Persistence/Models/Identity/Base/PermissionBaseFile.cs
@@ -1,6 +1,12 @@
 ﻿using Dayana.Shared.Persistence.Models.Identity.Base.Roles;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
-namespace Dayana.Shared.Persistence.Models.Identity.Base.Permissions;
+namespace Dayana.Shared.Persistence.Models.Identity.Base;
+
 
 public class PermissionModel
 {
