+++ Dayana/Shared/Persistence/Models/Identity/Base/RoleBaseFile.cs
@@ -1,0 +1,34 @@
+﻿using Dayana.Shared.Persistence.Models.Identity.Base.Roles;
+using Dayana.Shared.Persistence.Models.Identity.Base.Users;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Dayana.Shared.Persistence.Models.Identity.Base;
+public class RoleModel
+{
+    public int Id { get; set; }
+    public string Title { get; set; }
+
+    public DateTime CreatedAt { get; set; }
+    public DateTime UpdatedAt { get; set; }
+
+    public ICollection<UserRoleModel> UserRoles { get; set; }
+    public ICollection<RolePermissionModel> RolePermission { get; set; }
+    public List<PermissionModel> Permissions { get; set; }
+}
+
+
+public class RolePermissionModel
+{
+    public int RoleId { get; set; }
+    public int PermissionId { get; set; }
+
+    public DateTime CreatedAt { get; set; }
+    public DateTime UpdatedAt { get; set; }
+
+    public PermissionModel Permission { get; set; }
+    public RoleModel Role { get; set; }
+}
