--- Dayana/Shared/Persistence/Models/Identity/Base/Roles/RoleModel.cs
@@ -1,17 +1,0 @@
-﻿using Dayana.Shared.Persistence.Models.Identity.Base.Permissions;
-using Dayana.Shared.Persistence.Models.Identity.Base.Users;
-
-namespace Dayana.Shared.Persistence.Models.Identity.Base.Roles;
-
-public class RoleModel
-{
-    public int Id { get; set; }
-    public string Title { get; set; }
-
-    public DateTime CreatedAt { get; set; }
-    public DateTime UpdatedAt { get; set; }
-
-    public ICollection<UserRoleModel> UserRoles { get; set; }
-    public ICollection<RolePermissionModel> RolePermission { get; set; }
-    public List<PermissionModel> Permissions { get; set; }
-}
